remove dead method from uiapi and move to better folder
