+++ src/containers/graphics/views/fence.js
@@ -2,7 +2,7 @@ import Box2D from '../geometry/box2d';
 import invariant from 'invariant';
 import {
   clearSelection
-} from '../../uiapi';
+} from '../../../utils/ui/uiapi';
 
 /**
  * an interactive fence / drag box for the construct viewer
