+++ src/utils/ui/uiapi.js
@@ -1,0 +1,22 @@
+import invariant from 'invariant';
+
+/**
+ * clear any selected text in the entire document
+ */
+export function clearSelection() {
+
+  let selection = null;
+  if (window.getSelection) {
+    selection = window.getSelection();
+  } else if (document.selection) {
+    selection = document.selection;
+  }
+  if (selection) {
+    if (selection.empty) {
+      selection.empty();
+    }
+    if (selection.removeAllRanges) {
+      selection.removeAllRanges();
+    }
+  }
+}
